Edição dos metódos da reserva e a classe mesa

+++ Comanda.Api/Comanda.Api/Controllers/ReservasController.cs
@@ -7,6 +7,9 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Comanda.Api;
 using Comanda.Api.Models;
+using SQLitePCL;
+using Microsoft.AspNetCore.Http.Connections;
+using Microsoft.CodeAnalysis.Operations;
 
 namespace Comanda.Api.Controllers
 {
@@ -52,10 +55,35 @@ namespace Comanda.Api.Controllers
                 return BadRequest();
             }
 
+            // atualização
             _context.Entry(reserva).State = EntityState.Modified;
+
+            // mudar a situacao para livre da mesa original
+            // mudar a situacao para reservada da nova mesa
+            // remocao e inclusao da reserva na mesa (2 - reservada - 1.livre)
+            // 2 - livre 1 - reservada
+            // 
+            // ---------------
+            var novaMesa = await _context.Mesas.FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (novaMesa is null)
+                return BadRequest("Mesa não encontrada.");
+            novaMesa.SituacaoMesa = (int)SituacaoMesa.Reservado; // novaMesa agora está reservada
+
+            // mudar a situacao para livre da mesa original
+            var reservaOriginal = await _context.Reservas.AsNoTracking()
+                .FirstOrDefaultAsync(r => r.Id == id);
+
+            // consulta o numero da mesa original
+            var numeroMesaOriginal = reservaOriginal!.NumeroMesa;
+            // consulta a mesa original
+            var mesaOriginal = await _context.Mesas
+                .FirstOrDefaultAsync(m => m.NumeroMesa == numeroMesaOriginal);
+            mesaOriginal!.SituacaoMesa = (int)SituacaoMesa.Livre; // mesa original agora está livre
+            // ---------------------
 
             try
             {
+                // salva
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
@@ -79,6 +107,25 @@ namespace Comanda.Api.Controllers
         public async Task<ActionResult<Reserva>> PostReserva(Reserva reserva)
         {
             _context.Reservas.Add(reserva);
+
+            // Atualiza o status da mesa para "Reservado" 
+            var mesa = await _context.Mesas
+                .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            
+            if (mesa is null)
+                return BadRequest("Mesa não encontrada.");
+
+            // se mesa encontrada
+            if (mesa is not null)
+            {
+                if(mesa.SituacaoMesa != (int)SituacaoMesa.Livre)
+                {
+                    return BadRequest("Mesa não está disponível para reserva.");
+                }
+                // atualizar o status da mesa para reservado
+                mesa.SituacaoMesa = (int)SituacaoMesa.Reservado;
+            }
+            // -----------------
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetReserva", new { id = reserva.Id }, reserva);
@@ -93,7 +140,16 @@ namespace Comanda.Api.Controllers
             {
                 return NotFound();
             }
+            // -----------------------------
+            // consultar a mesa
+            var mesa = await _context.Mesas
+                .FirstOrDefaultAsync(m => m.NumeroMesa == reserva.NumeroMesa);
+            if (mesa is null)
+                return BadRequest("Mesa não encontrada.");
 
+            // atualizar a mesa para livre
+            mesa.SituacaoMesa = (int)SituacaoMesa.Livre; // (int) converte o enum para int
+            // -----------------------------
             _context.Reservas.Remove(reserva);
             await _context.SaveChangesAsync();
